requirements.txt
numpy
pandas 
matplotlib
seaborn 
Flask
scikit-learn
mysql-connector-python
# -e .